added options as state prop and modularized createBoilerFuncsCo-authored-by: Daniel Garan <g.daniel02@icloud.com> Co-authored-by: Kevin Can <kevincan321@gmail.com> Co-authored-by: Anthony Herrera <herreraanthony2@gmail.com> Co-authored-by: Rob Sinzieri <robsin@gmail.com> Co-authored-by: Roderick de Leon <roderickx.deleon@gmail.com>
