+++ src/components/nav-buttons/ExportMenu.vue
@@ -42,6 +42,7 @@ import {
   RouteComponentMap,
 } from "../../../types";
 import { useExportComponent } from "../composables/useExportComponent";
+import { createBoilerOptions, createBoilerComposition } from "../dfdas/createBoilerFuncs"
 // import * as fs from "fs"
 // import { fs } from "electron";
 // import { path } from 'path';
